frontend: split constants into separate PR
