+++ tests/filecheck/frontend/programs/programs.py
@@ -1,7 +1,5 @@
 # RUN: python %s | filecheck %s
 
-from xdsl.frontend.dialects.builtin import i32
-from xdsl.frontend.const import Const
 from xdsl.frontend.program import FrontendProgram
 from xdsl.frontend.context import CodeContext
 
@@ -9,18 +7,6 @@ p = FrontendProgram()
 with CodeContext(p):
     # CHECK: builtin.module() {}
     pass
-
-p.compile(desymref=False)
-print(p.xdsl())
-
-
-p = FrontendProgram()
-with CodeContext(p):
-
-    a: Const[i32] = 1
-
-    def foo():
-        pass
 
 p.compile(desymref=False)
 print(p.xdsl())
